Add security config and inmemory realization
Now program has two users - Roman and skipper. They are placed in memory and
if you need to add new User you will need to recompile the app.
It is good for testing. But it is unacceptable for production. So the next step
is to rewrite this code for multiplies users